safari-600.7-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.9_Automatic_Semicolon_Insertion/S7.9_A9_T7-expected.txt

CONSOLE MESSAGE: line 81: SyntaxError: Unexpected token '}'. Expected a statement as the body of a while loop.
S7.9_A9_T7

PASS Expected parsing failure

TEST COMPLETE
